fix:현재 주차에서 다음 주차 이동 불가기능 구현
