+++ src/pages/WeeklyDetail.jsx
@@ -5,26 +5,28 @@ import { useAuth } from '../context/Auth/AuthContext';
 import styled from 'styled-components';
 import KeyboardArrowLeftIcon from '@mui/icons-material/KeyboardArrowLeft';
 import KeyboardArrowRightIcon from '@mui/icons-material/KeyboardArrowRight';
-import {
-  getCurrentYearAndWeek,
-  getWeekAndFirstSundayDate,
-} from '../utils/dateUtil';
 import { useNavigate } from 'react-router-dom';
 import { BudgetRecipeSlide } from '../components/display/BudgetRecipeSlide';
 import { recommendAPI } from '../services/recommend.api';
+import { useWeeklyDate } from '../hooks/useWeeklyDate';
 
 // 날짜 계산 (헤더에 날짜 표시)
 const WeeklyDetail = () => {
   const navigate = useNavigate();
   const { state } = useAuth();
-  const [year, setYear] = useState(getCurrentYearAndWeek(new Date()).year);
-  const [week, setWeek] = useState(getCurrentYearAndWeek(new Date()).week);
-  const [currentMonth, setCurrentMonth] = useState('');
-  const [currentDate, setCurrentDate] = useState(new Date());
-  const [weekNumber, setWeekNumber] = useState(0);
-  const [firstSundayDateString, setFirstSundayDateString] = useState('');
   const [usedRecipes, setUsedRecipes] = useState([]); // 사용 레시피
   const [recommendedRecipes, setRecommendedRecipes] = useState([]); // 추천받은 레시피
+
+  // 날짜 관리 훅 가져오기
+  const {
+    year,
+    week,
+    currentMonth,
+    weekNumber,
+    isCurrentWeek,
+    handleDecreaseWeek,
+    handleIncreaseWeek,
+  } = useWeeklyDate();
 
   // 추천받은 레시피 정보 가져오기(is_used 무관)
   const getRecommendedRecipes = async () => {
@@ -55,45 +57,6 @@ const WeeklyDetail = () => {
     getUsedRecipes();
     getRecommendedRecipes();
   }, [year, week]);
-
-  useEffect(() => {
-    const { week, firstSundayDate } = getWeekAndFirstSundayDate(currentDate);
-    setWeekNumber(week);
-    setFirstSundayDateString(firstSundayDate.toLocaleDateString());
-
-    // 현재 월 설정
-    const monthNames = [
-      '1월',
-      '2월',
-      '3월',
-      '4월',
-      '5월',
-      '6월',
-      '7월',
-      '8월',
-      '9월',
-      '10월',
-      '11월',
-      '12월',
-    ];
-    setCurrentMonth(monthNames[firstSundayDate.getMonth()]); // 일요일 기준으로 현재 월을 설정
-  }, [currentDate]);
-
-  // 주차 변경 버튼 함수
-  const handleWeekChange = (delta) => {
-    const newDate = new Date(currentDate);
-    newDate.setDate(newDate.getDate() + delta * 7); // 1주일 더하기 또는 빼기
-    const { year: newYear, week: newWeek } = getCurrentYearAndWeek(newDate);
-    setUsedRecipes([]);
-    setRecommendedRecipes([]);
-    setCurrentDate(newDate);
-    setYear(newYear);
-    setWeek(newWeek);
-  };
-
-  // 주차 변경 버튼 클릭 핸들러
-  const handleDecreaseWeek = () => handleWeekChange(-1);
-  const handleIncreaseWeek = () => handleWeekChange(1);
 
   return (
     <Layout
@@ -148,13 +111,15 @@ const WeeklyDetail = () => {
             <RecipeListText>요리한 레시피가 없습니다.</RecipeListText>
             <br />
             <br />
-            <LinkButtonWrapper>
-              <LinkButton onClick={() => navigate('/home')}>
-                <p style={{ margin: '5px' }}>
-                  아직 요리한 레시피가 없네요! 요리하러 갈까요?
-                </p>
-              </LinkButton>
-            </LinkButtonWrapper>
+            {isCurrentWeek && ( // 현재 주차에서만 표시
+              <LinkButtonWrapper>
+                <LinkButton onClick={() => navigate('/home')}>
+                  <p style={{ margin: '5px' }}>
+                    아직 요리한 레시피가 없네요! 요리하러 갈까요?
+                  </p>
+                </LinkButton>
+              </LinkButtonWrapper>
+            )}
           </>
         )}
       </RecipeListBox>
